feat(rpc): split buy and sell rpc call
