+++ src/atomicdex/api/mm2/mm2.cpp
@@ -528,76 +528,6 @@ namespace mm2::api
     }
 
     void
-    from_json(const nlohmann::json& j, trading_order_contents& contents)
-    {
-        j.at("base").get_to(contents.base);
-        j.at("base_amount").get_to(contents.base_amount);
-        j.at("rel").get_to(contents.rel);
-        j.at("rel_amount").get_to(contents.rel_amount);
-        j.at("method").get_to(contents.method);
-        j.at("action").get_to(contents.action);
-        j.at("uuid").get_to(contents.uuid);
-        j.at("sender_pubkey").get_to(contents.sender_pubkey);
-        j.at("dest_pub_key").get_to(contents.dest_pub_key);
-    }
-
-    void
-    from_json(const nlohmann::json& j, buy_answer_success& contents)
-    {
-        j.get_to(contents.contents);
-    }
-
-    void
-    from_json(const nlohmann::json& j, buy_answer& answer)
-    {
-        if (j.count("error") == 1)
-        {
-            answer.error = j.at("error").get<std::string>();
-        }
-        else
-        {
-            answer.result = j.at("result").get<buy_answer_success>();
-        }
-    }
-
-    void
-    to_json(nlohmann::json& j, const buy_request& request)
-    {
-        j["base"]   = request.base;
-        j["price"]  = request.price;
-        j["rel"]    = request.rel;
-        j["volume"] = request.volume;
-        if (request.base_nota.has_value())
-        {
-            j["base_nota"] = request.base_nota.value();
-        }
-        if (request.base_confs.has_value())
-        {
-            j["base_confs"] = request.base_confs.value();
-        }
-        if (not request.is_created_order)
-        {
-            //! From orderbook
-            nlohmann::json price_fraction_repr  = nlohmann::json::object();
-            price_fraction_repr["numer"]        = request.price_numer;
-            price_fraction_repr["denom"]        = request.price_denom;
-            j["price"]                          = price_fraction_repr;
-            nlohmann::json volume_fraction_repr = nlohmann::json::object();
-            if (not request.selected_order_use_input_volume)
-            {
-                volume_fraction_repr["numer"] = request.volume_numer;
-                volume_fraction_repr["denom"] = request.volume_denom;
-                j["volume"]                   = volume_fraction_repr;
-            }
-            SPDLOG_INFO("The order is picked from the orderbook price: {}, volume: {}", j.at("price").dump(4), j.at("volume").dump(4));
-        }
-        else
-        {
-            SPDLOG_INFO("The order is not picked from orderbook we create it from volume = {}, price = {}", j.at("volume").dump(4), request.price);
-        }
-    }
-
-    void
     to_json(nlohmann::json& j, const setprice_request& request)
     {
         j["base"]            = request.base;
@@ -621,73 +551,6 @@ namespace mm2::api
         if (request.rel_confs.has_value())
         {
             j["rel_confs"] = request.rel_confs.value();
-        }
-    }
-
-    void
-    to_json(nlohmann::json& j, const sell_request& request)
-    {
-        SPDLOG_DEBUG("price: {}, volume: {}", request.price, request.volume);
-
-        auto volume_fraction_functor = [&request]() {
-            nlohmann::json volume_fraction_repr = nlohmann::json::object();
-            volume_fraction_repr["numer"]       = request.volume_numer;
-            volume_fraction_repr["denom"]       = request.volume_denom;
-            return volume_fraction_repr;
-        };
-
-        j["base"]   = request.base;
-        j["rel"]    = request.rel;
-        j["volume"] = request.volume; //< First take the user input
-        if (request.is_max)           //< It's a real max means user want to sell his base_max_taker_vol let's take the fraction repr
-        {
-            j["volume"] = volume_fraction_functor();
-        }
-        j["price"] = request.price;
-        if (request.rel_nota.has_value())
-        {
-            j["rel_nota"] = request.rel_nota.value();
-        }
-        if (request.rel_confs.has_value())
-        {
-            j["rel_confs"] = request.rel_confs.value();
-        }
-
-        if (not request.is_created_order)
-        {
-            //! From orderbook
-            nlohmann::json price_fraction_repr = nlohmann::json::object();
-            price_fraction_repr["numer"]       = request.price_numer;
-            price_fraction_repr["denom"]       = request.price_denom;
-            j["price"]                         = price_fraction_repr;
-            if (request.is_exact_selected_order_volume)
-            {
-                j["volume"] = volume_fraction_functor();
-            }
-            SPDLOG_INFO("The order is picked from the orderbook price: {}, volume: {}", j.at("price").dump(4), j.at("volume").dump(4));
-        }
-        else
-        {
-            SPDLOG_INFO("The order is not picked from orderbook we create it from volume = {}, price = {}", j.at("volume").dump(4), request.price);
-        }
-    }
-
-    void
-    from_json(const nlohmann::json& j, sell_answer_success& contents)
-    {
-        j.get_to(contents.contents);
-    }
-
-    void
-    from_json(const nlohmann::json& j, sell_answer& answer)
-    {
-        if (j.count("error") == 1)
-        {
-            answer.error = j.at("error").get<std::string>();
-        }
-        else
-        {
-            answer.result = j.at("result").get<sell_answer_success>();
         }
     }
